refactor(MetricsList): Ability to pass a variable name to the constructor
